# db/kursach/policeback/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Ревашольская милиция - OpenAPI 3.1.0
  version: 0.0.1
...
servers:
  - url: /api/v1
paths:
  /public/login:
    post:
      tags:
        - Аутентификация
      summary: Вход для полицейского
      description: После входа будет установлена cookie, которая позволит выполнять запросы от лица полицейского
      operationId: login
      requestBody: 
        description: Серийник и пароль
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/login'
      responses:
        '200':
          description: Аутентификация совершена, cookie установлена 
        '401':
          description: Если такого полицейского не существует или пароль оказался неверным. <br><br> Оба типа ошибки кодируются одинаково - это сделано намеренно, чтобы исключить выгрузку всех серийников злоумышленниками
...
  /auth/minor/note/create:
    post:
      tags:
        - Запись о деле
      summary: Создание заметки по делу
      operationId: noteCreate
      requestBody:
        description: Id дела и текст заметки
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/note-create'
      responses:
        '201':
          description: Успешно создана запись о деле
        '403':
          description: Полицейский не назначен на дело или снят с него
        '404':
          description: Такого преступления или дела на него нет
      security:
        - session:
          - minor
  /auth/major/policeman-case/put:
    put:
      tags:
        - Назначение полицейского на дело
      summary: Назначение начальником полицейского на открытое дело
      operationId: policemanCaseAdd
      requestBody:
        description: Id дела и полицейского, а также устанавливаемый статус (по умолчанию assigned)
        required: true
        content:
          applicaton/json:
            schema:
              $ref: '#/components/schemas/put-policeman-on-case'
      responses:
        '201':
          description: Полицейский успешно назначен на дело
        '204':
          description: Статус полицейского на деле успешно изменен
        '403':
          description: Полицейский не находится в подчинении
        '400':
          description: Дело не в статусе on_work или состояние на деле некорректное
        '404':
          description: Полицейского, преступления или дела не существует
      security:
        - session:
          - major
  /auth/minor/person-relevant-to-case/put:
    put:
      tags:
        - Участие человека в деле
      summary: Добавление / обновление человека как участвующего в деле
      requestBody:
        description: Id дела и человека, а также опциональные (если обновляем) статус его как участника и запись о нем
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/put-person-to-case'
      responses:
        '204':
          description: Участие человека в деле обновлено
        '201':
          description: Участие человека в деле отмечено
        '400':
          description: Дело не в статусе on_work или запись сделана о себе. Или если запись не существует (добавление) и статус или записка не указаны
        '403':
          description: Полицейский не назначен на дело или снят с него, а также (при обновлении) если он не владеет записью о человеке
        '404':
          description: Если дела или человека с указанным id не существует
      security:
        - session:
          - minor
...
  /auth/policeman/self:
    get:
      tags:
        - Полицейский
      summary: Получение основной информации о себе
      responses:
        '200':
          description: "Основная информация о себе: имя, звание, серийник, дата его выдачи"
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/base-policeman'
      security: 
        - session: [minor, major]
...
  /auth/major/slaves/my:
    get:
      tags:
        - Полицейский
      summary: Получение основной информации о подчиненных
      responses:
        '200':
          description: "Основная информация о подчиненных: id полицейского, имя, серийник, дата его выдачи"
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/no-context-no-rank-policemen'
      security: 
        - session: [major]
...
  /auth/minor/policeman/partners-in-case/{crime-id}:
    get:
      tags:
        - Полицейский
      summary: Получить основную информацию о партнерах по делу (кроме себя и снятых)
      parameters:
        - in: path
          name: crime-id
          required: true
          schema:
            type: integer
            example: 2
          description: id преступления, по которому ведется дело, по которому нужно получить основную информацию о партнерах
      responses:
        '200':
          description: "Основная информация о всех полицейских (кроме себя) на деле со статусом на деле"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/case-policemen'
        '404':
          description: Дела с таким id нет
        '403':
          description: Не назначен на дело или снят с него
      security: 
        - session: [minor]
  /auth/minor/note/in-case/{crime-id}:
    get:
      tags:
        - Запись о деле
      summary: Получить записи по делу
      parameters:
        - in: path
          name: crime-id
          required: true
          schema:
            type: integer
            example: 2
          description: id преступления, по которому ведется дело, по которому нужно получить записи
      responses:
        '200':
          description: "Записи по делу от партнеров (и себя): id, id полицейского, запись и список уникальных авторов записей (в формате основной информации)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/crime-notes'
        '404':
          description: Таких преступления или дела нет
        '403':
          description: Не назначен на дело или снят с него
      security: 
        - session: [minor]
  /auth/minor/person-relevant-to-case/in-case/{crime-id}:
    get:
      tags:
        - Участие человека в деле
      summary: Получить информацию о людях, участвующих в деле
      parameters:
        - in: path
          name: crime-id
          required: true
          schema:
            type: integer
            example: 2
          description: id преступления, по которому ведется дело, по которому нужно получить людей в нем участвующих
      responses:
        '200':
          description: "Информация о таких людях: id, имя, раса, отношение, запись о нем"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/base-case-people'
        '404':
          description: Такого преступления или дела нет
        '403':
          description: Не назначен на дело или снят с него
      security: 
        - session: [minor]
components:
  securitySchemes: 
    session:
      type: apiKey
      in: cookie
      name: JSESSIONID
      flows: 
        implicit: 
          scopes: 
            major: начальник, заведающий отделом
            minor: обычный полицейский, находящийся в подчинении начальника
      description: При выполнении входа добавляет в сессию информацию об этом, после этого в этой сессии можно выполнять запросы от своего имени
  schemas:
...
    person-relation:
      type: string
      enum:
        - witness
        - suspect
      example: witness
    policeman-case-status:
      type: string
      enum:
        - assigned
        - removed
      example: assigned
    race:
      type: string
      enum:
        - samaran
        - seolite
        - occidental
        - kojko
        - modial
        - perikarnassia
        - messina
        - oranje
        - mesque
        - vesper
        - sur-la-clef
        - ubi sent
        - graad
        - konigstein
        - zsiemsk
        - yugo-graad
        - igaunija
        - zemlya
        - mirova
        - ilmarran
        - semenine
        - revachol
        - ozonne
        - vaasa
        - suru
    rank:
      type: string
      enum:
        - major
        - minor
      description: major - это начальник minor полицейского
    login:
      type: object
      properties:
        series:
          type: string
        password:
          type: string
      examples:
        - series: 213e512d
          password: a21brf12d1aasd1
...
    note-create:
      type: object
      properties:
        crimeId:
          type: integer
        text:
          type: string
      examples:
        - crimeId: 3
          text: Нашли труп на Авенью стрит
    put-policeman-on-case:
      type: object
      properties:
        minorPolicemanId:
          type: integer
        crimeId:
          type: integer
        onCaseStatus:
          $ref: '#/components/schemas/policeman-case-status'
      examples:
        - minorPolicemanId: 3
          crimeId: 3
          onCaseStatus: removed
...
    put-person-to-case:
      type: object
      required: 
        - crimeId
        - personId
      properties:
        crimeId: 
          type: integer
        personId:
          type: integer
        relation:
          $ref: '#/components/schemas/person-relation'
        note:
          type: string
      examples:
        - crimeId: 3
          personId: 3
          relation: suspect
          note: Подозрительно много ходил у места преступления
...
    base-policeman:
      type: object
      properties:
        name:
          type: string
        rank:
          $ref: '#/components/schemas/rank'
        series:
          type: string
        seriesDateOfIssue:
          type: string
          format: date-time
      examples:
        - name: Валера
          rank: minor
          series: 123e512d
          seriesDateOfIssue: 2024-08-12T20:17:46.384Z
...
    no-context-no-rank-policeman:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        series:
          type: string
        seriesDateOfIssue:
          type: string
          format: date-time
      examples:
        - id: 29
          name: Валера
          series: 214e241d
          seriesDateOfIssue: '2024-08-12T20:17:46.384Z'
    no-context-no-rank-policemen:
      type: array
      items:
        $ref: '#/components/schemas/no-context-no-rank-policeman'
      examples:
        - - id: 29
            name: Валера
            series: 214e241d
            seriesDateOfIssue: '2024-08-12T20:17:46.384Z'
          - id: 32
            name: Алексей
            series: 327e925d
            seriesDateOfIssue: '2025-04-23T14:28:10.502Z'
          - id: 34
            name: Елена
            series: 518e232d
            seriesDateOfIssue: '2023-11-05T08:45:21.731Z'
...
    case-policeman:
      type: object
      properties:
        id: 
          type: integer
        name:
          type: string
        series:
          type: string
        seriesDateOfIssue:
          type: string
          format: date-time
        onCaseStatus:
          $ref: '#/components/schemas/policeman-case-status'
      examples:
        - id: 29
          name: Валера
          series: 214e241d
          seriesDateOfIssue: '2024-08-12T20:17:46.384Z'
          onCaseStatus: assigned
    case-policemen:
      type: array
      items:
          $ref: '#/components/schemas/case-policeman'
      examples:
        - - id: 29
            name: Валера
            series: 214e241d
            seriesDateOfIssue: '2024-08-12T20:17:46.384Z'
            onCaseStatus: assigned
          - id: 32
            name: Алексей
            series: 327e925d
            seriesDateOfIssue: '2025-04-23T14:28:10.502Z'
            onCaseStatus: removed
          - id: 34
            name: Елена
            series: 518e232d
            seriesDateOfIssue: '2023-11-05T08:45:21.731Z'
            onCaseStatus: assigned
    crime-note:
      type: object
      properties:
        id:
          type: integer
        policemanId:
          type: integer
        note:
          type: string
      examples:
        - id: 2
          policemanId: 29
          note: Найден труп на Авенью стрит
    crime-notes:
      type: array
      items:
          $ref: '#/components/schemas/crime-note'
      examples:
        - - id: 2
            policemanId: 29
            note: Найден труп на Авенью стрит
          - id: 5
            policemanId: 29
            note: Еще один труп с таким же подчерком на заводе
          - id: 8
            policemanId: 32
            note: У второго трупа видели человека в черной шляпе, он пытался скрыться
    base-case-person:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        race:
          $ref: '#/components/schemas/race'
        relation:
          $ref: '#/components/schemas/person-relation'
        note:
          type: string
      examples:
        - id: 4
          name: Анастасия
          race: messina
          relation: witness
          note: Проходила мимо места преступления - обнаружила первый труп
    base-case-people:
      type: array
      items:
        $ref: '#/components/schemas/base-case-person'
      examples:
        - - id: 4
            name: Анастасия
            race: messina
            relation: witness
            note: Проходила мимо места преступления - обнаружила первый труп
          - id: 7
            name: Илья
            race: oranje
            relation: witness
            note: Проходил мимо места преступления - обнаружил второй труп и описал подозрительно спешащего человека, уходящего от места преступления